\name{phrapl-internal}
\alias{Arrow3d2}
\alias{ColCountIf}
\alias{ColCountLast}
\alias{ColCountLastZero}
\alias{ColMax}
\alias{ColMin}
\alias{CombineSubsampleLikelihoods}
\alias{CompleteIntervals}
\alias{ConvertOutputVectorToLikelihood}
\alias{CreateAssignment.df}
\alias{CreateAssignment}
\alias{CreateMSstringGeneric}
\alias{CreateMSstringSpecific}
\alias{Expansionmultiplierindividual}
\alias{FractionNonZeroMigration}
\alias{GenerateExpansionMultiplierIndividuals}
\alias{GenerateIntervals}
\alias{GenerateIntervalsFullyResolvedCollapse}
\alias{GenerateIntervalsNoCollapse}
\alias{GenerateMigrationArrayMapTrunc}
\alias{GenerateMigrationIndividualsAllowNoMigration}
\alias{GenerateMigrationIndividualsFullyResolvedCollapseAllowNoMigration}
\alias{GenerateMigrationIndividualsNoCollapseAllowNoMigration}
\alias{GenerateN0multiplierIndividuals}
\alias{GetPermutationWeightsBasedOnSampling}
\alias{KAll}
\alias{KCollapseMatrix}
\alias{KN0multiplierMap}
\alias{KPopInterval}
\alias{LoadMS}
\alias{MatrixContainsAllValues}
\alias{Migrationindividual}
\alias{N0multiplierindividual}
\alias{PipeMS}
\alias{Popinterval}
\alias{PrunedAssignFrame}
\alias{PrunedPopVector}
\alias{ReturnAIC}
\alias{ReturnIncompletes}
\alias{ReturnModel}
\alias{ReturnSymmetricMigrationOnly}
\alias{RowMax}
\alias{RowMin}
\alias{SaveMS}
\alias{SearchContinuousModelSpaceNLoptr}
\alias{SearchContinuousModelSpaceRGenoud}
\alias{SetMaxK}
\alias{TaxaToDrop}
\alias{TaxaToRetain}
\alias{UpdateCompletes}
\alias{AllParamCombinations}
\alias{CalculateNullWeights}
\alias{CheckFiniteCoalescence}
\alias{ConvertOutputVectorToLikelihood.1sub}
\alias{CreateFineGrid}
\alias{CreateStartGrid}
\alias{ExtractAICs}
\alias{ExtractGridAICs}
\alias{ExtractGridParameters}
\alias{ExtractParameters}
\alias{FilterForFullyResolved}
\alias{GetAICweights}
\alias{GetAndBindLabel}
\alias{GetCladeNamesFromNewickString}
\alias{GetClades}
\alias{GetCladesQuickly}
\alias{GetKishESS}
\alias{GetOutDegreeOfPolytomies}
\alias{GetPermutationWeights}
\alias{GetScaledNreps}
\alias{GetScoreOfSingleTree}
\alias{GetTrianglePlotWeights}
\alias{NumberPopulationsAtRoot}
\alias{PassBounds}
\alias{ProportionDifference}
\alias{RemoveParameterNAs}
\alias{RowComplete}
\alias{SplitAndSort}
\alias{SubsampleMSTree}
\alias{SubsampleMSTree.raw}
\alias{SumDivScaledNreps}
\alias{SummarizeAveragedModelsAcrossSubsamples}
\alias{SummarizeParametersAcrossSubsamples}
\alias{TruncateToCells}
\alias{getExponent}
\alias{maxK}
\alias{AdjustUsingBeta}
\alias{ConvertTreeString}
\alias{FilterGridForMigration}
\alias{FilterMigrationByCriterion}
\alias{GenerateSetLoci}
\alias{GetLengthGridList}
\alias{GetNumTreeHistories}
\alias{GuessAssignments}
\alias{ScaleParameterVectorByNe}
\alias{TestForWithinPopCoalescenceThreeSamplesOnly}
\alias{DoSingleRowExtraction}
\alias{ExtractUnambiguousGridParameters}
\alias{ExtractUnambiguousNLoptrParameters}
\alias{ExtractUnambiguousOptimizationParameters}
\alias{ConcatenateResults_unambiguousParametersForOldRuns}
\alias{ConcatenateResults_ambiguousParameters}
\alias{sortParameterTable}
\alias{MsIndividualParametersConversionToUnambiguous}
\alias{RenumberWithoutGaps}
\alias{ModelAverageEachModel}
\alias{ExtractOptimizationAICs}
\alias{GenerateGrowthIndividuals}
\alias{GetCollapseGenerationNamesWhenAddedEventsExist}
\alias{KGrowthMap}
\alias{N0multiplierGrowthindividual}
\alias{PassBounds.addedEvents}
\alias{CompileMS}
\alias{CheckMS}
\title{Internal phrapl Functions}
\description{
  Internal \pkg{phrapl} functions.
}
\keyword{internal}
