\name{phrapl-package}
\alias{phrapl-package}
\alias{phrapl}
\docType{package}
\title{
Phylogeographic analysis using approximated likelihoods
}
\description{
  \pkg{phrapl} compares phylogeographic models.

  The complete list of functions can be displayed with
  \code{library(help = phrapl)}.

  More information on \pkg{phrapl} can be found at \url{http://phrapl-package.ird.fr/}.
}

\author{
  Brian O'Meara, Bryan Carstens, Nathan Jackson, Ariadna Morales-Garcia

  Maintainer: Brian O'Meara <bomeara@utk.edu>
}
\references{
  O'Meara, B.C., N. Jackson, A. Morales-Garcia, and B. Carstens (2014) Phrapl in prep.
}
\keyword{package}
