\name{phrapl-internal}
\alias{Arrow3d2}
\alias{ColCountIf}
\alias{ColCountLast}
\alias{ColMax}
\alias{ColMin}
\alias{CombineSubsampleLikelihoods}
\alias{CompleteIntervals}
\alias{ConvertOutputVectorToLikelihood}
\alias{CreateAssignment}
\alias{CreateAssignment.df}
\alias{CreateMSstringGeneric}
\alias{CreateMSstringSpecific}
\alias{ExhaustiveSearchNLoptr}
\alias{ExhaustiveSearchOptim}
\alias{Expansionmultiplierindividual}
\alias{GenerateExpansionMultiplierIndividuals}
\alias{GenerateIntervals}
\alias{GenerateIntervalsFullyResolvedCollapse}
\alias{GenerateIntervalsNoCollapse}
\alias{GenerateMigrationArrayMap}
\alias{GenerateMigrationIndividuals}
\alias{GenerateMigrationIndividualsAllowNoMigration}
\alias{GenerateMigrationIndividualsFullyResolvedCollapseAllowNoMigration}
\alias{GenerateMigrationIndividualsNoCollapseAllowNoMigration}
\alias{GenerateN0multiplierIndividuals}
\alias{KAll}
\alias{KCollapseMatrix}
\alias{KN0multiplierMap}
\alias{KPopInterval}
\alias{LoadMS}
\alias{MatrixContainsAllValues}
\alias{Migrationindividual}
\alias{MsIndividualParameters}
\alias{N0multiplierindividual}
\alias{PipeMS}
\alias{PlotModel}
\alias{Popinterval}
\alias{PrepSubsampling}
\alias{PrunedAssignFrame}
\alias{PrunedPopVector}
\alias{ReturnAIC}
\alias{ReturnIncompletes}
\alias{ReturnModel}
\alias{ReturnSymmetricMigrationOnly}
\alias{RowMax}
\alias{RowMin}
\alias{SaveMS}
\alias{SaveMovie}
\alias{SearchContinuousModelSpaceNLoptr}
\alias{SearchContinuousModelSpaceOptim}
\alias{SearchDiscreteModelSpaceNLoptr}
\alias{SearchDiscreteModelSpaceOptim}
\alias{SetMaxK}
\alias{TaxaToDrop}
\alias{TaxaToRetain}
\alias{UpdateCompletes}
\alias{AddDescriptors}
\alias{AllParamCombinations}
\alias{CalculateModelAverages}
\alias{CalculateNullWeights}
\alias{CheckFiniteCoalescence}
\alias{ConcatenateResults}
\alias{ConvertAlleleNumbersToPopulationLetters.phylo}
\alias{ConvertAlleleNumbersToPopulationLetters.raw}
\alias{ConvertOutputVectorToLikelihood.1sub}
\alias{CreateFineGrid}
\alias{CreateStartGrid}
\alias{ExtractAICs}
\alias{ExtractGridAICs}
\alias{ExtractGridParameters}
\alias{ExtractParameters}
\alias{FilterForFullyResolved}
\alias{GetAICweights}
\alias{GetAndBindLabel}
\alias{GetCladeNamesFromNewickString}
\alias{GetClades}
\alias{GetCladesQuickly}
\alias{GetKishESS}
\alias{GetOutDegreeOfPolytomies}
\alias{GetPermutatedLabels.raw}
\alias{GetPermutationWeights}
\alias{GetPermutationWeights.raw}
\alias{GetPermutationWeightsAcrossSubsamples}
\alias{GetPermutationWeightsAcrossSubsamples.raw}
\alias{GetScaledNreps}
\alias{GetScoreOfSingleTree}
\alias{GetTrianglePlotWeights}
\alias{MatchingTrees}
\alias{MergeTrees}
\alias{PassBounds}
\alias{ProportionDifference}
\alias{RemoveParameterNAs}
\alias{RowComplete}
\alias{RunRaxml}
\alias{RunSeqConverter}
\alias{SplitAndSort}
\alias{SubsampleMSTree.phylo}
\alias{SubsampleMSTree.raw}
\alias{SumDivScaledNreps}
\alias{SummarizeAveragedModelsAcrossSubsamples}
\alias{SummarizeParametersAcrossSubsamples}
\alias{TruncateToCells}
\alias{getExponent}
\alias{maxK}
\alias{migrationArray}
\title{Internal phrapl Functions}
\description{
  Internal \pkg{phrapl} functions.
}
\keyword{internal}
